Fix: change_line함수 수정
1. 달러는 무조건 환경변수로 변환
2. 작은따옴표는 문자로 취급
3. 큰 따옴표는 문자로 취급, 단 사이 문자열에 오로지 달러문자열만 있을
   경우 큰 따옴표삭제.

resolve #167
