+++ srcs/parsing/we_meet_other.c
@@ -6,12 +6,23 @@
 /*   By: sehjang <sehjang@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 18:58:10 by sehjang           #+#    #+#             */
-/*   Updated: 2022/10/05 00:49:15 by sunwchoi         ###   ########.fr       */
+/*   Updated: 2022/10/05 03:16:57 by sunwchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
-#include <stdio.h>
+
+void	change_first_node(t_list **first_node)
+{
+	t_list	*temp;
+
+	temp = *first_node;
+	*first_node = (*first_node)->next;
+	free(temp->content);
+	temp->content = NULL;
+	free(temp);
+	temp = NULL;
+}
 
 char	*we_meet_char(char *cmd, t_list **cmd_char_list)
 {
@@ -33,73 +44,27 @@ char	*we_meet_char(char *cmd, t_list **cmd_char_list)
 
 char	*we_meet_quotes(char *cmd, t_info *info, t_list **cmd_char_list)
 {
-	int		dollar_is_here;
 	t_list	*quotes;
+	int		only_dollar;
 
-	dollar_is_here = 0;
 	quotes = NULL;
 	cmd = we_meet_char(cmd, &quotes);
+	only_dollar = 0;
 	while (*cmd)
 	{
 		if (*cmd == '$')
 		{
 			cmd = we_meet_dollar(cmd, info, &quotes);
-			dollar_is_here++;
-		}
-		else if (*cmd == '"')
-		{
-			if (!dollar_is_here)
-				cmd = we_meet_char(cmd, &quotes);
-			break ;
+			only_dollar++;
 		}
 		else
-			cmd = we_meet_char(cmd, &quotes);
+			break ;
 	}
-	if (dollar_is_here && *cmd == '"')
+	if (*cmd == '"' && only_dollar)
 	{
 		change_first_node(&quotes);
 		cmd++;
 	}
 	ft_lstadd_back(cmd_char_list, quotes);
 	return (cmd);
-}
-
-char	*we_meet_dollar(char *cmd, t_info *info, t_list **cmd_char_list)
-{
-	t_list	*dollar;
-
-	dollar = NULL;
-	cmd = we_meet_char(cmd, &dollar);
-	if (*cmd == '?')
-		cmd = we_meet_char(cmd, &dollar);
-	else
-	{
-		while (*cmd)
-		{
-			if (ft_isalnum(*cmd) || *cmd == '_')
-				cmd = we_meet_char(cmd, &dollar);
-			else
-				break ;
-		}
-	}
-	change_node(&dollar, info);
-	ft_lstadd_back(cmd_char_list, dollar);
-	return (cmd);
-}
-
-t_list	*char2list(char *cmd, t_info *info)
-{
-	t_list	*cmd_char_list;
-
-	cmd_char_list = NULL;
-	while (*cmd)
-	{
-		if (*cmd == '$')
-			cmd = we_meet_dollar(cmd, info, &cmd_char_list);
-		else if (*cmd == '"')
-			cmd = we_meet_quotes(cmd, info, &cmd_char_list);
-		else
-			cmd = we_meet_char(cmd, &cmd_char_list);
-	}
-	return (cmd_char_list);
 }
